requirements.txt
django>=5.1.7
djangorestframework>=3.15.2
requests>=2.32.3
beautifulsoup4>=4.13.3
transformers>=4.49.0
torch>=2.6.0
gunicorn>=21.2.0
whitenoise>=6.6.0 